Start creating a menu and a recyclerView

+++ app/src/main/java/com/example/myapplication/data/entities/Task.kt
@@ -18,7 +18,7 @@ data class Task(
     @ColumnInfo(name = "text")
     var text: String,
     @ColumnInfo(name = "timestamp")
-    var timestamp: Int,
+    var timestamp: Long,
     @ColumnInfo(name = "state")
     var isDode: Boolean,
 )
